Moved makeHTTPClientBootstrapBase back to Utils.swift
It is now a member of NIOClientTCPBootstrap
